backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
gunicorn==21.2.0
google-generativeai==0.5.4
numpy==1.26.4
faiss-cpu==1.7.4
python-dotenv==1.0.1
PyMuPDF==1.24.1
python-docx==1.1.2  